Fix user id

+++ entities/favorites/src/Models/FavoriteModel.php
@@ -77,7 +77,7 @@ class FavoriteModel extends Model implements FavoriteModelContract
      */
     public function setUserIdAttribute($value)
     {
-        $this->attributes['user_id'] = (int) trim(strip_tags($value));
+        $this->attributes['user_id'] = trim(strip_tags($value));
     }
 
     /**
